+++ springboot-samples/web/webflux/base/src/main/java/com/alipay/sofa/web/webflux/DemoWebfluxApplication.java
@@ -16,14 +16,15 @@
  */
 package com.alipay.sofa.web.webflux;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootTest
-class DemowebfluxApplicationTests {
+@SpringBootApplication
+public class DemoWebfluxApplication {
 
-    @Test
-    void contextLoads() {
+    public static void main(String[] args) {
+        SpringApplication.run(DemoWebfluxApplication.class, args);
+        System.out.println("DemowebfluxApplication start!!!");
     }
 
 }
